+++ library/Zend/Search/Lucene/Index/SegmentWriter/AbstractSegmentWriter.php
@@ -114,7 +114,7 @@ abstract class AbstractSegmentWriter
      * normVector is a binary string.
      * Each byte corresponds to an indexed document in a segment and
      * encodes normalization factor (float value, encoded by
-     * \Zend\Search\Lucene\Search\Similarity::encodeNorm())
+     * \Zend\Search\Lucene\Search\Similarity\AbstractSimilarity::encodeNorm())
      *
      * @var array
      */
