+++ library/Zend/Search/Lucene/Search/Similarity/DefaultSimilarity.php
@@ -21,7 +21,7 @@
 
 namespace Zend\Search\Lucene\Search\Similarity;
 
-use Zend\Search\Lucene\Search\Similarity;
+use Zend\Search\Lucene\Search\Similarity\AbstractSimilarity;
 
 /**
  * @category   Zend
@@ -30,7 +30,7 @@ use Zend\Search\Lucene\Search\Similarity;
  * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
-class DefaultSimilarity extends Similarity
+class DefaultSimilarity extends AbstractSimilarity
 {
 
     /**
